move all controllers to separate namespace
